Split ApplyDeltaTests into multiple types
