+++ test/dotnet-watch.Tests/HotReload/SourceFileUpdateTests.cs
@@ -1,0 +1,364 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#nullable disable
+
+namespace Microsoft.DotNet.Watch.UnitTests
+{
+    public class SourceFileUpdateTests(ITestOutputHelper logger) : DotNetWatchTestBase(logger)
+    {
+        [Fact]
+        public async Task AddSourceFile()
+        {
+            Log("AddSourceFile started");
+
+            var testAsset = TestAssets.CopyTestAsset("WatchAppWithProjectDeps")
+                .WithSource();
+
+            var dependencyDir = Path.Combine(testAsset.Path, "Dependency");
+
+            App.Start(testAsset, [], "AppWithDeps");
+
+            await App.WaitUntilOutputContains(MessageDescriptor.WaitingForChanges);
+
+            // add a new file:
+            UpdateSourceFile(Path.Combine(dependencyDir, "AnotherLib.cs"), """
+                public class AnotherLib
+                {
+                    public static void Print()
+                        => System.Console.WriteLine("Changed!");
+                }
+                """);
+
+            await App.WaitUntilOutputContains(MessageDescriptor.ReEvaluationCompleted);
+
+            // update existing file:
+            UpdateSourceFile(Path.Combine(dependencyDir, "Foo.cs"), """
+                public class Lib
+                {
+                    public static void Print()
+                        => AnotherLib.Print();
+                }
+                """);
+
+            await App.WaitUntilOutputContains("Changed!");
+        }
+
+        [Fact]
+        public async Task ChangeFileInDependency()
+        {
+            var testAsset = TestAssets.CopyTestAsset("WatchAppWithProjectDeps")
+                .WithSource();
+
+            var dependencyDir = Path.Combine(testAsset.Path, "Dependency");
+
+            App.Start(testAsset, [], "AppWithDeps");
+
+            await App.WaitUntilOutputContains(MessageDescriptor.WaitingForChanges);
+
+            var newSrc = """
+                public class Lib
+                {
+                    public static void Print()
+                        => System.Console.WriteLine("Changed!");
+                }
+                """;
+
+            UpdateSourceFile(Path.Combine(dependencyDir, "Foo.cs"), newSrc);
+
+            await App.WaitUntilOutputContains("dotnet watch 🔥 Hot reload capabilities: AddExplicitInterfaceImplementation AddFieldRva AddInstanceFieldToExistingType AddMethodToExistingType AddStaticFieldToExistingType Baseline ChangeCustomAttributes GenericAddFieldToExistingType GenericAddMethodToExistingType GenericUpdateMethod NewTypeDefinition UpdateParameters.");
+
+            await App.WaitUntilOutputContains("Changed!");
+        }
+
+        /// <summary>
+        /// Unchanged project doesn't build. Wait for source change and rebuild.
+        /// </summary>
+        [Fact]
+        public async Task BaselineCompilationError()
+        {
+            var testAsset = TestAssets.CopyTestAsset("WatchNoDepsApp")
+                .WithSource();
+
+            var programPath = Path.Combine(testAsset.Path, "Program.cs");
+            File.WriteAllText(programPath,
+                """
+                Console.Write
+                """);
+
+            App.Start(testAsset, []);
+
+            await App.WaitUntilOutputContains(MessageDescriptor.WaitingForFileChangeBeforeRestarting);
+
+            UpdateSourceFile(programPath, """
+                System.Console.WriteLine("<Updated>");
+                """);
+
+            await App.WaitUntilOutputContains("<Updated>");
+        }
+
+        [Theory]
+        [InlineData("PublishAot", "True")]
+        [InlineData("PublishTrimmed", "True")]
+        [InlineData("StartupHookSupport", "False")]
+        public async Task ChangeFileInAotProject(string propertyName, string propertyValue)
+        {
+            var tfm = ToolsetInfo.CurrentTargetFramework;
+
+            var testAsset = TestAssets.CopyTestAsset("WatchHotReloadApp", identifier: $"{propertyName};{propertyValue}")
+                .WithSource()
+                .WithProjectChanges(project =>
+                {
+                    project.Root.Descendants()
+                        .First(e => e.Name.LocalName == "PropertyGroup")
+                        .Add(XElement.Parse($"<{propertyName}>{propertyValue}</{propertyName}>"));
+                });
+
+            var programPath = Path.Combine(testAsset.Path, "Program.cs");
+
+            App.Start(testAsset, ["--non-interactive"]);
+
+            await App.WaitForOutputLineContaining($"[WatchHotReloadApp ({tfm})] " + MessageDescriptor.ProjectDoesNotSupportHotReload.GetMessage($"'{propertyName}' property is '{propertyValue}'"));
+            await App.WaitForOutputLineContaining(MessageDescriptor.WaitingForChanges);
+            App.Process.ClearOutput();
+
+            UpdateSourceFile(programPath, content => content.Replace("Console.WriteLine(\".\");", "Console.WriteLine(\"<updated>\");"));
+
+            await App.WaitForOutputLineContaining($"[auto-restart] {programPath}(1,1): error ENC0097"); //  Applying source changes while the application is running is not supported by the runtime.
+            await App.WaitForOutputLineContaining("<updated>");
+        }
+
+        [Fact]
+        public async Task ChangeFileInFSharpProject()
+        {
+            var testAsset = TestAssets.CopyTestAsset("FSharpTestAppSimple")
+                .WithSource();
+
+            App.Start(testAsset, []);
+
+            await App.WaitUntilOutputContains(MessageDescriptor.WaitingForFileChangeBeforeRestarting);
+
+            UpdateSourceFile(Path.Combine(testAsset.Path, "Program.fs"), content => content.Replace("Hello World!", "<Updated>"));
+
+            await App.WaitUntilOutputContains("<Updated>");
+        }
+
+        [Fact]
+        public async Task ChangeFileInFSharpProjectWithLoop()
+        {
+            var testAsset = TestAssets.CopyTestAsset("FSharpTestAppSimple")
+                .WithSource();
+
+            var source = """
+            module ConsoleApplication.Program
+
+            open System
+            open System.Threading
+
+            [<EntryPoint>]
+            let main argv =
+                printfn "Waiting"
+                Thread.Sleep(Timeout.Infinite)
+                0
+            """;
+
+            var sourcePath = Path.Combine(testAsset.Path, "Program.fs");
+
+            File.WriteAllText(sourcePath, source);
+
+            App.Start(testAsset, ["--non-interactive"]);
+
+            await App.WaitUntilOutputContains(MessageDescriptor.WaitingForChanges);
+            App.Process.ClearOutput();
+
+            UpdateSourceFile(sourcePath, content => content.Replace("Waiting", "<Updated>"));
+
+            await App.WaitUntilOutputContains(MessageDescriptor.WaitingForChanges);
+            await App.WaitUntilOutputContains("<Updated>");
+            App.Process.ClearOutput();
+
+            UpdateSourceFile(sourcePath, content => content.Replace("<Updated>", "<Updated2>"));
+
+            await App.WaitUntilOutputContains(MessageDescriptor.WaitingForChanges);
+            await App.WaitUntilOutputContains("<Updated2>");
+        }
+
+        // Test is timing out on .NET Framework: https://github.com/dotnet/sdk/issues/41669
+        [CoreMSBuildOnlyFact]
+        public async Task HandleTypeLoadFailure()
+        {
+            var testAsset = TestAssets.CopyTestAsset("WatchAppTypeLoadFailure")
+                .WithSource();
+
+            App.Start(testAsset, [], "App");
+
+            await App.WaitUntilOutputContains(MessageDescriptor.WaitingForChanges);
+
+            var newSrc = """
+                class DepSubType : Dep
+                {
+                    int F() => 2;
+                }
+
+                class Printer
+                {
+                    public static void Print()
+                    {
+                        Console.WriteLine("Changed!");
+                    }
+                }
+                """;
+
+            UpdateSourceFile(Path.Combine(testAsset.Path, "App", "Update.cs"), newSrc);
+
+            await App.WaitUntilOutputContains("Updated types: Printer");
+        }
+
+        // Test is timing out on .NET Framework: https://github.com/dotnet/sdk/issues/41669
+        [CoreMSBuildOnlyFact]
+        public async Task HandleMissingAssemblyFailure()
+        {
+            var testAsset = TestAssets.CopyTestAsset("WatchAppMissingAssemblyFailure")
+                .WithSource();
+
+            App.Start(testAsset, [], "App");
+
+            await App.WaitUntilOutputContains(MessageDescriptor.WaitingForChanges);
+
+            var newSrc = /* lang=c#-test */"""
+                using System;
+
+                public class DepType
+                {
+                    int F() => 1;
+                }
+
+                public class Printer
+                {
+                    public static void Print()
+                        => Console.WriteLine("Updated!");
+                }
+                """;
+
+            // Delete all files in testAsset.Path named Dep.dll
+            foreach (var depDll in Directory.GetFiles(testAsset.Path, "Dep2.dll", SearchOption.AllDirectories))
+            {
+                File.Delete(depDll);
+            }
+
+            File.WriteAllText(Path.Combine(testAsset.Path, "App", "Update.cs"), newSrc);
+
+            await App.WaitUntilOutputContains("Updated types: Printer");
+        }
+
+        [Theory]
+        [InlineData(true, Skip = "https://github.com/dotnet/sdk/issues/43320")]
+        [InlineData(false)]
+        public async Task RenameSourceFile(bool useMove)
+        {
+            Log("RenameSourceFile started");
+
+            var testAsset = TestAssets.CopyTestAsset("WatchAppWithProjectDeps")
+                .WithSource();
+
+            var dependencyDir = Path.Combine(testAsset.Path, "Dependency");
+            var oldFilePath = Path.Combine(dependencyDir, "Foo.cs");
+            var newFilePath = Path.Combine(dependencyDir, "Renamed.cs");
+
+            var source = """
+                using System;
+                using System.IO;
+                using System.Runtime.CompilerServices;
+
+                public class Lib
+                {
+                    public static void Print() => PrintFileName();
+
+                    public static void PrintFileName([CallerFilePathAttribute] string filePath = null)
+                    {
+                        Console.WriteLine($"> {Path.GetFileName(filePath)}");
+                    }
+                }
+                """;
+
+            File.WriteAllText(oldFilePath, source);
+
+            App.Start(testAsset, [], "AppWithDeps");
+
+            await App.WaitUntilOutputContains(MessageDescriptor.WaitingForChanges);
+
+            // rename the file:
+            if (useMove)
+            {
+                File.Move(oldFilePath, newFilePath);
+            }
+            else
+            {
+                File.Delete(oldFilePath);
+                File.WriteAllText(newFilePath, source);
+            }
+
+            Log($"Renamed '{oldFilePath}' to '{newFilePath}'.");
+
+            await App.AssertOutputLineStartsWith("> Renamed.cs");
+        }
+
+        [Theory]
+        [InlineData(true, Skip = "https://github.com/dotnet/sdk/issues/43320")]
+        [InlineData(false)]
+        public async Task RenameDirectory(bool useMove)
+        {
+            Log("RenameSourceFile started");
+
+            var testAsset = TestAssets.CopyTestAsset("WatchAppWithProjectDeps", identifier: useMove.ToString())
+                .WithSource();
+
+            var dependencyDir = Path.Combine(testAsset.Path, "Dependency");
+            var oldSubdir = Path.Combine(dependencyDir, "Subdir");
+            var newSubdir = Path.Combine(dependencyDir, "NewSubdir");
+
+            var source = """
+                using System;
+                using System.IO;
+                using System.Runtime.CompilerServices;
+
+                public class Lib
+                {
+                    public static void Print() => PrintDirectoryName();
+
+                    public static void PrintDirectoryName([CallerFilePathAttribute] string filePath = null)
+                    {
+                        Console.WriteLine($"> {Path.GetFileName(Path.GetDirectoryName(filePath))}");
+                    }
+                }
+                """;
+
+            File.Delete(Path.Combine(dependencyDir, "Foo.cs"));
+            Directory.CreateDirectory(oldSubdir);
+            File.WriteAllText(Path.Combine(oldSubdir, "Foo.cs"), source);
+
+            App.Start(testAsset, ["--non-interactive"], "AppWithDeps");
+
+            await App.WaitUntilOutputContains(MessageDescriptor.WaitingForChanges);
+
+            // rename the directory:
+            if (useMove)
+            {
+                Directory.Move(oldSubdir, newSubdir);
+            }
+            else
+            {
+                Directory.Delete(oldSubdir, recursive: true);
+                Directory.CreateDirectory(newSubdir);
+                File.WriteAllText(Path.Combine(newSubdir, "Foo.cs"), source);
+            }
+
+            Log($"Renamed '{oldSubdir}' to '{newSubdir}'.");
+
+            // dotnet-watch may observe the delete separately from the new file write.
+            // If so, rude edit is reported, the app is auto-restarted and we should observe the final result.
+
+            await App.WaitUntilOutputContains("> NewSubdir");
+        }
+    }
+}
